Don't warn repeatedly for same un-optimize-able Op
Closes #23.
